requirement.txt
python==3.11.5
django==4.2
djangorestframework==3.14.0
python-dotenv==1.0.0
mysqlclient==2.2.0
django-cors-headers==4.6.0

